account tools for user and group management (incl. testing)

+++ planlos/account_tools.py
@@ -3,7 +3,8 @@
 from planlos.passwordgen import Password_Generator
 from flaskext.mail import Message
 from planlos.extensions import mail, db
-from documents import User, Group, Group_Admin, User_Exists_Exception
+from documents import User, Group, Group_Admin
+from documents import User_Exists_Exception, Group_Exists
 from flask import current_app as app
 from datetime import datetime
 import uuid
@@ -20,8 +21,24 @@ def find_all_group_members(group):
     pass
 
 
+def create_group(groupname):
+    if db.Group.find_one({'name': groupname}):
+        raise Group_Exists
+    g = db.Group()
+    g.name = groupname
+    g.desc = u''
+    g.save()
+
+
+def remove_group(groupname):
+    group = db.Group.find_one({'name': groupname})
+    if group is None:
+        raise LookupError
+    group.delete()
+
 def create_user(username, password, email):
-    if db.User.find_one({'username': username}) == None and db.Group.find_one({'name': username}) == None:
+    if (db.User.find_one({'username': username}) == None
+        and db.Group.find_one({'name': username}) == None):
         # create group
         g = db.Group()
         g.name = username
@@ -34,7 +51,7 @@ def create_user(username, password, email):
         user.password = password
         user.signup_email = email
         user.activation_key = generate_activation_code()
-        user.groups = [ g ]
+        user.groups = [g]
         user.save()
 
         ga = db.Group_Admin()
@@ -47,18 +64,37 @@ def create_user(username, password, email):
         raise User_Exists_Exception()
 
 
+def group_add(group, user):
+    user = db.User.find_one({'username': user})
+    if user is None:
+        raise LookupError
+    user.groups.remove(group)
+    user.save()
+
+
+def remove_user_from_group(user, group):
+    user = db.User.find_one({'username': user})
+    if user is None:
+        raise LookupError
+    user.groups.remove(group)
+    user.save()
+
+
 def remove_user(username):
     user = db.User.find_one({'username': username})
+    if user is None:
+        raise LookupError
     group = db.Group.find_one({'name': username})
     groupadmin = db.Group_Admin.find_one({'group': group._id})
     groupadmin.delete()
+
+    # remove user from groups
+    #nicht notwendig
+    # remove other groups if empty
+
+    # remove user
     user.delete()
     group.delete()
-
-
-    # remove other groups if empty
-    # remove user
-    # send email?
     # signal admin dashboard
